README
To use the database, just open it with the command "use data"; "select * from general"; both commands open the table and show its contents.
The table shows the 6 days of operation, Sunday being a day that the company is closed. On Saturday it is customary to outline and publish weekly tasks, as Saturday is the Home Office day, the day everyone is on call and allows easier communication with the entire company.